requirements.txt
# Core
pytz==2023.3
Django==4.2.1
django-configurations==2.4.1
gunicorn==20.1.0
newrelic==8.8.0
python-decouple==3.8

# For the persistence stores
psycopg2-binary==2.9.6
dj-database-url==2.0.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers==4.2.0

# Documentation
drf-spectacular==0.26.4

# Developer Tools
ipdb==0.13.13
ipython==8.13.2
mkdocs==1.4.3
flake8==6.0.0

# Testing
mock==5.0.2
factory-boy==3.2.1
django-nose==1.4.7
coverage==7.2.5

# Static and Media Storage
cloudinary==1.32.0
django-cloudinary-storage==0.3.0
whitenoise==6.5.0
Pillow==10.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
